slider 메뉴 이벤트 drag로 변경 / 모드 셀렉터 화면 크기에 맞춰 비율 조정

+++ Assets/SimulFactory/Script/Game/Event/S_Login.cs
@@ -14,7 +14,7 @@ namespace SimulFactory.Game.Event
             {
                 message.Add("userNo", param[1]);
             }
-            EventManager.GetInstance().TriggerEvent((byte)Define.UNITY_EVENT.Login, message);
+            EventManager.GetInstance().TriggerEvent(Define.UNITY_EVENT.Login, message);
         }
     }
 }
